Changed eventId output to float.

+++ users/marcel/avgraph/vfxNodes/vfxNodeTimeline.h
@@ -61,7 +61,7 @@ struct VfxNodeTimeline : VfxNodeBase
 	
 	VfxTimeline * timeline;
 	
-	int eventIdOutput;
+	float eventIdOutput;
 
 	VfxNodeTimeline();
 	virtual ~VfxNodeTimeline() override;
